requirements.txt
beautifulsoup4==4.13.4
langchain==0.3.26
langchain_core==0.3.69
langchain_openai==0.3.28
python-dotenv==1.1.1
Requests==2.32.4
selenium==4.34.2
streamlit==1.46.1
supabase==2.17.0
webdriver_manager==4.0.2
